CNN-Mnist/requirements.txt
torch==1.10.0
torchvision==0.11.1
numpy==1.21.4
pyyaml==6.0.2
tqdm==4.61.2